♻️ Refactor images, extract group no data comp.
